+++ packages/polymath-issuer/src/pages/sto/components/ConfigureSTOForm/forms/USDTieredSTO/index.js
@@ -5,9 +5,8 @@ import { map, reduce } from 'lodash';
 import { FastField, withFormik } from 'formik';
 import moment from 'moment-timezone';
 import { Form, Tooltip, Button } from 'carbon-components-react';
-import Web3 from 'web3';
 import BigNumber from 'bignumber.js';
-import validator from '@polymathnetwork/shared/validator';
+import validator from '@polymathnetwork/ui/validator';
 import {
   Box,
   Grid,
@@ -22,6 +21,16 @@ import {
   NumberInput,
   CurrencySelect,
 } from '@polymathnetwork/ui';
+import {
+  validateEndDate,
+  validateEndTime,
+  validateTodayOrAfter,
+  validateStartTime,
+  REQUIRED_MESSAGE,
+  MIN_MESSAGE,
+  MORE_THAN_MESSAGE,
+  ADDRESS_MESSAGE,
+} from '../../validators';
 import InvestmentTiers from './InvestmentTiers';
 import { toWei } from '../../../../../../utils/contracts';
 import { FUND_RAISE_TYPES } from '../../../../../../constants';
@@ -30,8 +39,6 @@ import { configureSTO } from '../../../../../../actions/sto';
 import type { Dispatch } from 'redux';
 import type { RootState } from '../../../../../../redux/reducer';
 import type { FundRaiseType } from '../../../../../../constants';
-
-const TRANSACTION_TIME_BUFFER = 20 * 60 * 1000;
 
 type InvestmentTier = {|
   tokensAmount: number,
@@ -73,88 +80,15 @@ type ContainerProps = {|
   ...FormikProps,
 |};
 
-function validateEndTime(value) {
-  const startDate: Date = this.parent.startDate;
-  const startTime: number = this.parent.startTime;
-  const endDate: Date = this.parent.endDate;
-  if (!startDate || !startTime || !endDate) {
-    return true;
-  }
-  const startUnix = moment(startDate).unix() * 1000 + startTime;
-  const endUnix = moment(endDate).unix() * 1000 + value;
-  if (startUnix >= endUnix) {
-    return this.createError({ message: 'End time must be after start time.' });
-  }
-
-  return true;
-}
-
-function validateEndDate(value) {
-  const startDate: Date = this.parent.startDate;
-  const valid = moment(value).isSameOrAfter(startDate);
-
-  if (!valid) {
-    return this.createError({ message: 'End date must be after start date.' });
-  }
-  return true;
-}
-
-function todayOrAfter(value) {
-  const valid = moment(value).isSameOrAfter(moment(Date.now()).startOf('day'));
-  if (valid) {
-    return true;
-  }
-
-  return this.createError({ message: 'Start time must be today or later.' });
-}
-
-function validateStartTime(value) {
-  const startDate: Date = this.parent.startDate;
-
-  if (!startDate) {
-    return true;
-  }
-
-  const startUnix = moment(startDate).unix() * 1000 + value;
-  const nowUnix = moment(Date.now()).unix() * 1000;
-  const timeUntilStart = startUnix - nowUnix;
-
-  if (nowUnix >= startUnix) {
-    return this.createError({ message: 'Start time is in the past.' });
-  }
-  if (timeUntilStart < TRANSACTION_TIME_BUFFER) {
-    return this.createError({
-      message: 'Please allow for transaction processing time.',
-    });
-  }
-  return true;
-}
-
-function validateIsAddress(value) {
-  if (!Web3.utils.isAddress(value)) {
-    return this.createError({
-      message: 'Is an invalid address',
-    });
-  }
-
-  return true;
-}
-
-const requiredMessage = 'Required.';
-/* eslint-disable no-template-curly-in-string */
-const moreThanMessage = 'Must be higher than ${more}.';
-const minMessage = 'Must be at least ${min}.';
-/* eslint-enable no-template-curly-in-string */
-
 export const investmentTierSchema = validator.object().shape({
   tokensAmount: validator
     .bigNumber()
-    .isRequired(requiredMessage)
-    .moreThan(0, moreThanMessage),
+    .isRequired(REQUIRED_MESSAGE)
+    .moreThan(0, MORE_THAN_MESSAGE),
   tokenPrice: validator
     .bigNumber()
-    .isRequired(requiredMessage)
-    .moreThan(0, moreThanMessage),
+    .isRequired(REQUIRED_MESSAGE)
+    .moreThan(0, MORE_THAN_MESSAGE),
 });
 
 // TODO @RafaelVidaurre: Move reusable validators to yup singleton
@@ -162,41 +96,41 @@ const formSchema = validator.object().shape({
   date: validator.object().shape({
     startDate: validator
       .date()
-      .isRequired(requiredMessage)
-      .test('validateStartDate', todayOrAfter),
+      .isRequired(REQUIRED_MESSAGE)
+      .test('validateStartDate', validateTodayOrAfter),
     startTime: validator
       .number()
-      .isRequired(requiredMessage)
+      .isRequired(REQUIRED_MESSAGE)
       .test('validateStartTime', validateStartTime),
     endDate: validator
       .date()
-      .isRequired(requiredMessage)
+      .isRequired(REQUIRED_MESSAGE)
       .test('validateEndDate', validateEndDate),
     endTime: validator
       .number()
-      .isRequired(requiredMessage)
+      .isRequired(REQUIRED_MESSAGE)
       .test('validEndTime', validateEndTime),
   }),
   currencies: validator
     .array()
     .of(validator.string())
-    .isRequired(requiredMessage),
+    .isRequired(REQUIRED_MESSAGE),
   nonAccreditedMax: validator
     .bigNumber()
-    .isRequired(requiredMessage)
-    .min(0, minMessage),
+    .isRequired(REQUIRED_MESSAGE)
+    .min(0, MIN_MESSAGE),
   minimumInvestment: validator
     .bigNumber()
-    .isRequired(requiredMessage)
-    .min(0, minMessage),
+    .isRequired(REQUIRED_MESSAGE)
+    .min(0, MIN_MESSAGE),
   receiverAddress: validator
     .string()
-    .isRequired(requiredMessage)
-    .test('validateIsAddress', validateIsAddress),
+    .isRequired(REQUIRED_MESSAGE)
+    .isAddress(ADDRESS_MESSAGE),
   unsoldTokensAddress: validator
     .string()
-    .isRequired(requiredMessage)
-    .test('validateIsAddress', validateIsAddress),
+    .isRequired(REQUIRED_MESSAGE)
+    .isAddress(ADDRESS_MESSAGE),
   investmentTiers: validator.object().shape({
     isMultipleTiers: validator.boolean(),
     tiers: validator
@@ -445,7 +379,7 @@ const formikEnhancer = withFormik({
     return initialValues;
   },
   handleSubmit: (values: FormValues, { props }) => {
-    const { dispatch, address } = props;
+    const { dispatch } = props;
 
     const formattedValues = {
       startsAt:
@@ -475,7 +409,7 @@ const formikEnhancer = withFormik({
       data: formattedValues,
     };
 
-    dispatch(configureSTO(address, config)).catch(error => {
+    dispatch(configureSTO(config, 'USDTieredSTO')).catch(error => {
       throw error;
     });
   },
